+++ src/Instance/InstanceDetails/SubInstanceList/SubInstanceList.js
@@ -1,17 +1,24 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Link } from 'react-router-dom';
+import { CopyToClipboard } from 'react-copy-to-clipboard';
 
 import {
   MultiColumnList,
-  KeyValue,
   NoValue,
+  IconButton,
 } from '@folio/stripes/components';
 
-import { getIdentifiers } from '../../../utils';
+import {
+  getIdentifiers,
+  formatCellStyles,
+} from '../../../utils';
 import { indentifierTypeNames } from '../../../constants';
 import useReferenceData from '../../../hooks/useReferenceData';
+import useCallout from '../../../hooks/useCallout';
+
+import css from './SubInstanceList.css';
 
 const noValue = <NoValue />;
 
@@ -28,6 +35,13 @@ const SubInstanceList = ({
   const {
     identifierTypesById,
   } = useReferenceData();
+  const callout = useCallout();
+  const onCopyToClipbaord = useCallback(hrid => {
+    callout.sendCallout({
+      type: 'success',
+      message: <FormattedMessage id="ui-inventory.hridCopied" values={{ hrid }} />,
+    });
+  }, [callout]);
 
   const formatter = {
     title: row => (row[titleKey] ?
@@ -37,7 +51,19 @@ const SubInstanceList = ({
         {row.title}
       </Link> :
       row.title || noValue),
-    hrid: row => row.hrid || noValue,
+    hrid: row => (
+      row.hrid ?
+        <>
+          {row.hrid}
+          <CopyToClipboard
+            text={row.hrid}
+            onCopy={() => onCopyToClipbaord(row.hrid)}
+          >
+            <IconButton icon="clipboard" />
+          </CopyToClipboard>
+        </> :
+        noValue
+    ),
     publisher: row => row.publication?.[0]?.publisher ?? noValue,
     publisherDate: row => row.publication?.[0]?.dateOfPublication ?? noValue,
     issn: row => getIdentifiers(row.identifiers, ISSN, identifierTypesById) || noValue,
@@ -64,22 +90,21 @@ const SubInstanceList = ({
 
   const columnWidths = {
     title: '40%',
-    hrid: '25%',
+    hrid: '30%',
   };
 
   return (
-    <KeyValue label={label}>
-      <MultiColumnList
-        id={id}
-        contentData={titles}
-        visibleColumns={visibleColumns}
-        columnMapping={columnMapping}
-        columnWidths={columnWidths}
-        formatter={formatter}
-        ariaLabel={label}
-        interactive={false}
-      />
-    </KeyValue>
+    <MultiColumnList
+      id={id}
+      contentData={titles}
+      visibleColumns={visibleColumns}
+      columnMapping={columnMapping}
+      columnWidths={columnWidths}
+      formatter={formatter}
+      getCellClass={formatCellStyles(css.cellAlign)}
+      ariaLabel={label}
+      interactive={false}
+    />
   );
 };