+++ src/ViewHoldingsRecord.js
@@ -36,6 +36,7 @@ import {
 import {
   AppIcon,
   IntlConsumer,
+  CalloutContext,
 } from '@folio/stripes/core';
 
 import {
@@ -55,8 +56,13 @@ import {
   noValue,
   holdingsStatementTypes,
 } from './constants';
+import { WarningMessage } from './components';
+
+import css from './View.css';
 
 class ViewHoldingsRecord extends React.Component {
+  static contextType = CalloutContext;
+
   static manifest = Object.freeze({
     query: {},
     permanentLocationQuery: {},
@@ -144,6 +150,13 @@ class ViewHoldingsRecord extends React.Component {
     if (holdings.temporaryLocationId === '') delete holdings.temporaryLocationId;
 
     return this.props.mutator.holdingsRecords.PUT(holdings).then(() => {
+      this.context.sendCallout({
+        type: 'success',
+        message: <FormattedMessage
+          id="ui-inventory.holdingsRecord.successfullySaved"
+          values={{ hrid: holdingsRecord.hrid }}
+        />,
+      });
       this.onClickCloseEditHoldingsRecord();
     });
   }
@@ -528,14 +541,22 @@ class ViewHoldingsRecord extends React.Component {
                   </Row>
                   <hr />
                   <AccordionStatus>
-                    <Row>
-                      <Col xs={10}>
-                        <MessageBanner show={holdingsRecord.discoverySuppress} type="warning">
-                          <FormattedMessage id="ui-inventory.warning.holdingsRecord.suppressedFromDiscovery" />
-                        </MessageBanner>
+                    <Row className={css.rowMarginBottom}>
+                      <Col xs={11}>
+                        <Row center="xs" middle="xs">
+                          <Col>
+                            <MessageBanner show={Boolean(holdingsRecord.discoverySuppress)} type="warning">
+                              <FormattedMessage id="ui-inventory.warning.holdingsRecord.suppressedFromDiscovery" />
+                            </MessageBanner>
+                          </Col>
+                        </Row>
                       </Col>
-                      <Col data-test-expand-all xs={2}>
-                        <ExpandAllButton />
+                      <Col data-test-expand-all xs={1}>
+                        <Row end="xs">
+                          <Col>
+                            <ExpandAllButton />
+                          </Col>
+                        </Row>
                       </Col>
                     </Row>
                     <AccordionSet initialStatus={initialAccordionsState}>
@@ -544,6 +565,11 @@ class ViewHoldingsRecord extends React.Component {
                         label={<FormattedMessage id="ui-inventory.administrativeData" />}
                       >
                         <this.cViewMetaData metadata={holdingsRecord.metadata} />
+                        <Row>
+                          <Col xs={12}>
+                            {holdingsRecord.discoverySuppress && <WarningMessage id="ui-inventory.discoverySuppressed" />}
+                          </Col>
+                        </Row>
                         <br />
                         <Row>
                           <Col