+++ src/ViewInstance.js
@@ -2,9 +2,10 @@ import {
   get,
 } from 'lodash';
 import React, {
-  createRef,
+  createRef
 } from 'react';
 import PropTypes from 'prop-types';
+import { parse } from 'query-string';
 import ReactRouterPropTypes from 'react-router-prop-types';
 import { FormattedMessage } from 'react-intl';
 
@@ -12,6 +13,7 @@ import {
   AppIcon,
   IfPermission,
   IfInterface,
+  Pluggable,
 } from '@folio/stripes/core';
 import {
   Pane,
@@ -29,12 +31,15 @@ import {
   batchFetchItems,
   batchFetchRequests,
 } from './Instance/ViewRequests/utils';
+import { indentifierTypeNames, layers } from './constants';
+import { DataContext } from './contexts';
 
 import {
   HoldingsListContainer,
   MoveItemsContext,
   InstanceDetails,
 } from './Instance';
+import { CalloutRenderer } from './components';
 
 import ImportRecordModal from './components/ImportRecordModal';
 
@@ -113,6 +118,8 @@ class ViewInstance extends React.Component {
       findInstancePluginOpened: false,
       isItemsMovement: false,
       isImportRecordModalOpened: false,
+      isCopyrightModalOpened: false,
+      afterCreate: false,
     };
     this.instanceId = null;
     this.cViewHoldingsRecord = this.props.stripes.connect(ViewHoldingsRecord);
@@ -138,6 +145,13 @@ class ViewInstance extends React.Component {
 
     if (isMARCSource && instanceRecordsId !== prevInstanceRecordsId) {
       this.getMARCRecord();
+    }
+
+    // component got updated after a new record was created
+    if (parse(prevProps?.location?.search)?.layer === layers.CREATE &&
+      !parse(this.props?.location?.search)?.layer && !this.state.afterCreate) {
+      // eslint-disable-next-line
+      this.setState({ afterCreate: true });
     }
 
     const { allInstanceHoldings, allInstanceItems } = resources;
@@ -289,8 +303,39 @@ class ViewInstance extends React.Component {
     this.setState({ isImportRecordModalOpened: false });
   }
 
+  toggleCopyrightModal = () => {
+    this.setState(prevState => ({ isCopyrightModalOpened: !prevState.isCopyrightModalOpened }));
+  };
+
   toggleFindInstancePlugin = () => {
     this.setState(prevState => ({ findInstancePluginOpened: !prevState.findInstancePluginOpened }));
+  };
+
+  // Get all identifiers for all records
+  getIdentifiers = (data) => {
+    const { identifierTypesById } = data;
+    const { ISBN, ISSN } = indentifierTypeNames;
+    const records = this.props?.resources?.selectedInstance?.records;
+    if (!records) {
+      return null;
+    }
+    // We can't make any meaningful assessment of which is
+    // the best identifier to return, so just return the first
+    // we find
+    for (const record of records) {
+      for (const identifiers of record.identifiers) {
+        const { identifierTypeId, value } = identifiers;
+        const ident = identifierTypesById[identifierTypeId];
+        if (
+          (ident?.name === ISBN ||
+            ident?.name === ISSN) &&
+          value
+        ) {
+          return { type: ident.name, value };
+        }
+      }
+    }
+    return null;
   };
 
   createActionMenuGetter = instance => ({ onToggle }) => {
@@ -463,6 +508,31 @@ class ViewInstance extends React.Component {
             </Button>
           </IfPermission>
         </IfInterface>
+        <DataContext.Consumer>
+          {data => (
+            <Pluggable
+              id="copyright-permissions-checker"
+              toggle={this.toggleCopyrightModal}
+              open={this.state.isCopyrightModalOpened}
+              identifier={this.getIdentifiers(data)}
+              type="copyright-permissions-checker"
+              renderTrigger={({ menuText }) => (
+                <Button
+                  id="copyright-permissions-check"
+                  buttonStyle="dropdownItem"
+                  onClick={() => {
+                    onToggle();
+                    this.toggleCopyrightModal();
+                  }}
+                >
+                  <Icon icon="report">
+                    {menuText}
+                  </Icon>
+                </Button>
+              )}
+            />
+          )}
+        </DataContext.Consumer>
       </>
     );
   };
@@ -538,6 +608,12 @@ class ViewInstance extends React.Component {
         </InstanceDetails>
 
         <Callout ref={this.calloutRef} />
+
+        {this.state.afterCreate &&
+          <CalloutRenderer
+            message={<FormattedMessage id="ui-inventory.instance.successfullySaved" values={{ hrid: instance.hrid }} />}
+          />
+        }
 
         {
           this.state.findInstancePluginOpened && (